Changes done to fix the error in iPage server.

+++ sql/sql.sql
@@ -5,7 +5,7 @@ code varchar(25),
 email varchar(40),
 confirmed boolean,
 colorVestido varchar(50),
-tallaPie decimal(2,2),
+tallaPie float,
 message varchar(500),
 PRIMARY KEY (ID)
 );
